+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changes
 =======
 
+In next release ...
+
+Bugfixes:
+
+- An attributes in i18n:attributes no longer needs to match an
+  existing or dynamic attribute in order to appear in the
+  element. This fixes issue #66.
+
 2.5.3 (2011-10-23)
 ------------------